+++ ui/src/views/ProfileShow.vue
@@ -45,7 +45,7 @@
       v-if="dialog === 'edit-community'"
       :show="dialog === 'edit-community'"
       editing
-      :title="t('addPersonFormTitle', { displayName: profile.preferredName })"
+      :title="dt('addPersonFormTitle', { displayName: profile.preferredName })"
       @delete="dialog = 'delete-community'"
       @submit="updateCommunity"
       @close="dialog = null"
@@ -54,7 +54,7 @@
     <EditNodeDialog
       v-if="dialog === 'edit-node'"
       :show="dialog === 'edit-node'"
-      :title="t('addPersonFormTitle', { displayName: getDisplayName(profile) })"
+      :title="dt('addPersonFormTitle', { displayName: getDisplayName(profile) })"
       @submit="updatePerson"
       @close="dialog = null"
       :profile="profile"
@@ -209,7 +209,7 @@ export default {
       var name = ''
       if (this.profile.legalName) name = this.profile.legalName
       else if (this.profile.preferredName) name = this.profile.preferredName
-      if (name.length > 30) return 'font-size:6vw'
+      if (name.length > 30) return 'font-size:6vw; text-align:center; margin-left:15px; margin-right:15px'
       if (name.length > 25) return 'font-size:7vw'
       if (name.length > 20) return 'font-size:8vw'
       else return 'font-size: 10vw'
@@ -332,6 +332,9 @@ export default {
     },
     t (key, vars) {
       return this.$t('viewPerson.' + key, vars)
+    },
+    dt (key, vars) {
+      return this.$t('addPersonForm.' + key, vars)
     }
   }
 }