+++ desktop/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog | Ahau desktop
 
-## Current
+## 1.5.0
 
 ### Features
 
@@ -14,6 +14,7 @@
 - Updated Pataka onboarding
   - updated the Pataka dialog to provided users with other options
   - Users will now be prompted to join a Pātaka upon creating a profile 
+- Translation updates
 
 
 ### Bug Fixes
@@ -29,6 +30,9 @@
 - fix addPersonFormTitle translations
 - Fixed bug causing selected locale to default to English when another language has been selected
 - Updated the popup messages for consistency in style
+- bugfix around public profiles
+    - now impossible to accidentally publish private profile (e.g. phone number) to your public (unencrypted) profile) over graphql
+    - ensure all public profiles have do not have private profile data in them
 
 ## v1.4.2